# personal-secretary/openapi.yaml
openapi: 3.1.0
info:
  title: Gurukul Personal Secretary
  version: 1.2.0
  description: >-
    Private quote-and-confirm action platform for Harsh. It preserves the Del
    Ray Barbershop haircut workflow, a sandbox bakery delivery workflow, and a
    budget-capped monthly floral subscription checkout handoff. Every
    consequential action requires an unchanged short-lived signed quote token
    and explicit confirmation.
servers:
  - url: https://gurukul-personal-secretary-harsh.onrender.com
paths:
  /health:
    get:
      operationId: checkSecretaryHealth
      summary: Check whether the secretary service is running and configured.
      x-openai-isConsequential: false
      security: []
      responses:
        '200':
          description: Health status.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  configured:
                    type: boolean
                  service:
                    type: string
                  workflows:
                    type: array
                    items:
                      type: string
  /v1/availability:
    post:
      operationId: findUsualHaircutAvailability
      summary: Find the earliest approved live haircut appointment.
      description: >-
        Searches Del Ray Barbershop for Buzzcut and Beard Trim with Scotty.
        Thursday is preferred over Friday within each searched week, and only
        times at or after 2:00 PM America/Vancouver are eligible. The search is
        limited to the next two weeks so it completes within the Action timeout.
      x-openai-isConsequential: false
      security:
        - SecretaryApiKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                max_weeks:
                  type: integer
                  minimum: 1
                  maximum: 2
                  default: 2
                  description: Search one or two upcoming weeks.
      responses:
        '200':
          description: A live appointment quote.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AvailabilityQuote'
        '404':
          description: No matching availability was found.
        '401':
          description: Invalid API key.
  /v1/book:
    post:
      operationId: bookUsualHaircut
      summary: Book the exact haircut appointment contained in a live quote.
      description: >-
        This creates an external appointment. Call only with the unmodified
        quote token returned by findUsualHaircutAvailability and only after the
        user has approved that exact date and time. No payment is charged.
      x-openai-isConsequential: true
      security:
        - SecretaryApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmQuoteRequest'
      responses:
        '200':
          description: Confirmed appointment.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BookingConfirmation'
        '409':
          description: The quoted slot is no longer available.
        '400':
          description: Confirmation or quote validation failed.
        '401':
          description: Invalid API key.
  /v1/bakery/delivery-quote:
    post:
      operationId: getBakeryDeliveryQuote
      summary: Get a sandbox bakery delivery quote.
      description: >-
        Creates a proposal using only configured merchant, product, delivery,
        quantity, tip, tax, and spending constraints. This endpoint uses a fake
        provider and cannot place a real order or charge payment.
      x-openai-isConsequential: false
      security:
        - SecretaryApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BakeryDeliveryQuoteRequest'
      responses:
        '200':
          description: Sandbox bakery delivery quote.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BakeryDeliveryQuote'
        '400':
          description: Request validation failed.
        '401':
          description: Invalid API key.
        '502':
          description: Bakery quote provider failed.
  /v1/bakery/order:
    post:
      operationId: placeBakeryDeliveryOrder
      summary: Confirm the exact sandbox bakery quote.
      description: >-
        Executes only the unmodified quote returned by getBakeryDeliveryQuote
        and only with explicit confirmation. This is sandbox-only in the current
        implementation and cannot place a real bakery order.
      x-openai-isConsequential: true
      security:
        - SecretaryApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmQuoteRequest'
      responses:
        '200':
          description: Sandbox bakery order result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BakeryOrderResult'
        '400':
          description: Confirmation or quote validation failed.
        '401':
          description: Invalid API key.
        '502':
          description: Sandbox bakery provider failed.
  /v1/flowers/subscription-quote:
    post:
      operationId: getMonthlyFlowerSubscriptionQuote
      summary: Quote the approved monthly floral subscription.
      description: >-
        Checks the V&J public catalog for the Classic Floral Subscription,
        three-month variant, and monthly selling plan. The price must remain at
        or below CAD 80 per delivery. No address, payment data, order, or
        subscription is created by this operation.
      x-openai-isConsequential: false
      security:
        - SecretaryApiKey: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FlowerSubscriptionQuoteRequest'
      responses:
        '200':
          description: Live florist catalog quote.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowerSubscriptionQuote'
        '400':
          description: Request validation failed.
        '409':
          description: The configured plan is unavailable or over budget.
        '401':
          description: Invalid API key.
        '502':
          description: Florist catalog lookup failed.
  /v1/flowers/prepare-checkout:
    post:
      operationId: prepareMonthlyFlowerSubscriptionCheckout
      summary: Recheck and return the approved florist checkout handoff.
      description: >-
        Rechecks the exact signed quote and returns the official product page
        plus the pinned checkout selections. It does not place an order, start
        a subscription, transmit a delivery address, or charge payment. The
        user must review and complete checkout with the florist.
      x-openai-isConsequential: true
      security:
        - SecretaryApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmQuoteRequest'
      responses:
        '200':
          description: Verified checkout handoff.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FlowerCheckoutResult'
        '400':
          description: Confirmation or quote validation failed.
        '409':
          description: Florist availability, price, or terms changed.
        '401':
          description: Invalid API key.
        '502':
          description: Florist catalog recheck failed.
components:
  securitySchemes:
    SecretaryApiKey:
      type: apiKey
      in: header
      name: X-Secretary-Key
  schemas:
    ConfirmQuoteRequest:
      type: object
      additionalProperties: false
      required:
        - quote_token
        - confirm
      properties:
        quote_token:
          type: string
          description: Unmodified short-lived quote token from the matching quote endpoint.
        confirm:
          type: boolean
          enum: [true]
          description: Must be true after user confirmation.
    Slot:
      type: object
      required:
        - dateLabel
        - timeLabel
        - business
        - service
        - staff
        - location
        - estimatedPriceCad
        - dueTodayCad
      properties:
        dateLabel:
          type: string
          examples: ['Friday, Aug 7, 2026']
        timeLabel:
          type: string
          examples: ['2:40 p.m.']
        business:
          type: string
          const: Del Ray Barbershop
        service:
          type: string
          const: Buzzcut and Beard Trim
        staff:
          type: string
          const: Scotty
        location:
          type: string
        estimatedPriceCad:
          type: number
        dueTodayCad:
          type: number
          const: 0
    AvailabilityQuote:
      type: object
      required:
        - slot
        - quote_token
        - quote_expires_at
        - requires_confirmation
      properties:
        slot:
          $ref: '#/components/schemas/Slot'
        quote_token:
          type: string
        quote_expires_at:
          type: string
          format: date-time
        requires_confirmation:
          type: boolean
          const: true
    BookingConfirmation:
      type: object
      required:
        - confirmed
        - business
        - service
        - staff
        - date
        - time
        - location
      properties:
        confirmed:
          type: boolean
          const: true
        idempotent_replay:
          type: boolean
        business:
          type: string
        service:
          type: string
        staff:
          type: string
        date:
          type: string
        time:
          type: string
        location:
          type: string
        estimatedDueAtAppointmentCad:
          type: number
    BakeryDeliveryQuoteRequest:
      type: object
      additionalProperties: false
      required:
        - products
        - delivery_destination_ref
      properties:
        merchant_id:
          type: string
          const: sandbox-bakery
          default: sandbox-bakery
        products:
          type: array
          minItems: 1
          maxItems: 3
          items:
            $ref: '#/components/schemas/BakeryProductRequest'
        delivery_destination_ref:
          type: string
          enum: [home, office]
          description: Configured destination reference only. Do not send raw addresses.
        delivery_window:
          type: string
          enum: [asap, today_4_6_pm, tomorrow_9_11_am]
          default: asap
        tip_cad:
          type: number
          minimum: 0
          maximum: 10
          default: 0
    BakeryProductRequest:
      type: object
      additionalProperties: false
      required:
        - sku
        - quantity
      properties:
        sku:
          type: string
          enum:
            - sourdough-loaf
            - croissant-box-6
            - cinnamon-roll-box-4
        quantity:
          type: integer
          minimum: 1
          maximum: 4
    BakeryDeliveryQuote:
      type: object
      required:
        - workflow
        - proposal
        - quote_token
        - quote_expires_at
        - requires_confirmation
      properties:
        workflow:
          type: string
          const: bakery_delivery
        proposal:
          $ref: '#/components/schemas/BakeryProposal'
        quote_token:
          type: string
        quote_expires_at:
          type: string
          format: date-time
        requires_confirmation:
          type: boolean
          const: true
    BakeryProposal:
      type: object
      required:
        - workflow
        - provider
        - live_order
        - merchant
        - products
        - delivery_destination_ref
        - delivery_window
        - fees
        - subtotal_cad
        - tax_cad
        - tip_cad
        - total_estimated_cad
        - currency
      properties:
        workflow:
          type: string
          const: bakery_delivery
        provider:
          type: string
          const: fake_bakery
        live_order:
          type: boolean
          const: false
        merchant:
          type: object
          required: [id, name]
          properties:
            id:
              type: string
              const: sandbox-bakery
            name:
              type: string
              const: Sandbox Bakery
        products:
          type: array
          items:
            $ref: '#/components/schemas/BakeryQuotedProduct'
        delivery_destination_ref:
          type: string
          enum: [home, office]
        delivery_window:
          $ref: '#/components/schemas/BakeryDeliveryWindow'
        fees:
          type: object
          required: [delivery_cad]
          properties:
            delivery_cad:
              type: number
        subtotal_cad:
          type: number
        tax_cad:
          type: number
        tip_cad:
          type: number
        total_estimated_cad:
          type: number
        currency:
          type: string
          const: CAD
    BakeryQuotedProduct:
      type: object
      required:
        - sku
        - name
        - quantity
        - unit_price_cad
        - line_total_cad
      properties:
        sku:
          type: string
        name:
          type: string
        quantity:
          type: integer
        unit_price_cad:
          type: number
        line_total_cad:
          type: number
    BakeryDeliveryWindow:
      type: object
      required: [id, label]
      properties:
        id:
          type: string
        label:
          type: string
    BakeryOrderResult:
      type: object
      required:
        - workflow
        - order_id
        - status
        - live_order
        - merchant
        - products
        - delivery_destination_ref
        - delivery_window
        - total_cad
        - currency
      properties:
        idempotent_replay:
          type: boolean
        workflow:
          type: string
          const: bakery_delivery
        order_id:
          type: string
        status:
          type: string
          const: accepted_sandbox
        live_order:
          type: boolean
          const: false
        merchant:
          type: object
        products:
          type: array
          items:
            $ref: '#/components/schemas/BakeryQuotedProduct'
        delivery_destination_ref:
          type: string
        delivery_window:
          $ref: '#/components/schemas/BakeryDeliveryWindow'
        total_cad:
          type: number
        currency:
          type: string
          const: CAD
        provider_reference:
          type: string
          const: fake_bakery
    FlowerSubscriptionQuoteRequest:
      type: object
      additionalProperties: false
      properties:
        delivery_destination_ref:
          type: string
          enum: [wife_home]
          default: wife_home
          description: Configured destination reference only. Never send a raw address.
    FlowerSubscriptionQuote:
      type: object
      required:
        - workflow
        - proposal
        - quote_token
        - quote_expires_at
        - requires_confirmation
      properties:
        workflow:
          type: string
          const: flower_subscription
        proposal:
          $ref: '#/components/schemas/FlowerSubscriptionProposal'
        quote_token:
          type: string
        quote_expires_at:
          type: string
          format: date-time
        requires_confirmation:
          type: boolean
          const: true
    FlowerSubscriptionProposal:
      type: object
      required:
        - workflow
        - provider
        - live_order
        - merchant
        - product
        - subscription
        - delivery_destination_ref
        - cadence
        - commitment_deliveries
        - price_per_delivery_cad
        - maximum_per_delivery_cad
        - currency
        - requires_checkout
        - checkout_url
      properties:
        workflow:
          type: string
          const: flower_subscription
        provider:
          type: string
          const: vj_shopify_checkout
        live_order:
          type: boolean
          const: false
        merchant:
          $ref: '#/components/schemas/FlowerMerchant'
        product:
          $ref: '#/components/schemas/FlowerProduct'
        subscription:
          $ref: '#/components/schemas/FlowerSubscriptionTerms'
        delivery_destination_ref:
          type: string
          const: wife_home
        cadence:
          type: string
          const: monthly
        commitment_deliveries:
          type: integer
          const: 3
        price_per_delivery_cad:
          type: number
          maximum: 80
        maximum_per_delivery_cad:
          type: number
          const: 80
        currency:
          type: string
          const: CAD
        requires_checkout:
          type: boolean
          const: true
        checkout_url:
          type: string
          format: uri
    FlowerMerchant:
      type: object
      required: [id, name, location]
      properties:
        id:
          type: string
          const: vj-plant-shop
        name:
          type: string
          const: V&J Plant Shop
        location:
          type: string
    FlowerProduct:
      type: object
      required: [id, title, variant_id, variant_title]
      properties:
        id:
          type: integer
        title:
          type: string
          const: Classic Floral Subscription
        variant_id:
          type: integer
        variant_title:
          type: string
          const: 3 Months
    FlowerSubscriptionTerms:
      type: object
      required: [selling_plan_id, name, frequency]
      properties:
        selling_plan_id:
          type: integer
        name:
          type: string
          const: Monthly (3 Month)
        frequency:
          type: string
          const: monthly
    FlowerCheckoutResult:
      type: object
      required:
        - workflow
        - status
        - live_order
        - merchant
        - product
        - subscription
        - delivery_destination_ref
        - commitment_deliveries
        - price_per_delivery_cad
        - maximum_per_delivery_cad
        - currency
        - checkout_url
        - checkout_instructions
      properties:
        idempotent_replay:
          type: boolean
        workflow:
          type: string
          const: flower_subscription
        status:
          type: string
          const: checkout_required
        live_order:
          type: boolean
          const: false
        merchant:
          $ref: '#/components/schemas/FlowerMerchant'
        product:
          $ref: '#/components/schemas/FlowerProduct'
        subscription:
          $ref: '#/components/schemas/FlowerSubscriptionTerms'
        delivery_destination_ref:
          type: string
          const: wife_home
        commitment_deliveries:
          type: integer
          const: 3
        price_per_delivery_cad:
          type: number
          maximum: 80
        maximum_per_delivery_cad:
          type: number
          const: 80
        currency:
          type: string
          const: CAD
        checkout_url:
          type: string
          format: uri
        checkout_instructions:
          type: string